// frontend/weather-monitor/lib/types.ts
export interface WeatherData {
  city: string;
  country: string;
  latitude: number;
  longitude: number;
  temperature: number;
  apparent_temperature: number;
  humidity: number;
  precipitation: number;
  rain: number;
  wind_speed: number;
  wind_direction: number;
  cloud_cover: number;
  pressure: number;
  weather_code: number;
  timestamp: string;
}